CHORE: 주석 추가 및 import 수정
